+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Spark 4.0.0
 * Java 11 compile target
+* Updated docs and code snippets for `staticParams` interceptor
+    * Confluence 8.0 introduced `staticParams` as a replacement for `static-params`, the latter of which has been removed in Confluence 9.0. 
 
 # Spark 3.1.0
 * Introduced support for dark mode, by running `SPARK.initializeTheming()` in the parent window.